After some waffling, I think for the moment I'll try putting buffers in the `FastBandlimited` object, so that `mul!` and friends can be (i) non-allocating and (ii) more broadly applicable. `finufft_exec!` has strict type bounds, so doing this and copying things over into compliant buffers will mean that you can `mul!` with many more things.
The challenge here is that plans are made for specific numbers of
columns in advance. So with this new pre-computed plan object in the
struct, that means that if you have ten columns but your plan is only
computed for one, you don't get the nice speedup from concurrently doing
multiple RHS vectors at once. Still thinking about how to handle that.